Fix audit log dependencies
