+++ source/dotnet/wholesale-api/Calculations/Calculations.Application/AuditLog/AuditLogger.cs
@@ -12,9 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Energinet.DataHub.Core.App.Common.Abstractions.Users;
 using Energinet.DataHub.Core.JsonSerialization;
 using Energinet.DataHub.Core.Outbox.Abstractions;
 using Energinet.DataHub.Wholesale.Calculations.Interfaces.AuditLog;
+using Energinet.DataHub.Wholesale.Common.Interfaces.Security;
 using NodaTime;
 
 namespace Energinet.DataHub.Wholesale.Calculations.Application.AuditLog;
@@ -24,7 +26,7 @@ public class AuditLogger(
     IClock clock,
     IOutboxClient outboxClient,
     IUnitOfWork unitOfWork,
-    IAuditUserContext auditUserContext)
+    IUserContext<FrontendUser> userContext)
     : IAuditLogger
 {
     private static readonly Guid _wholesaleSystemId = Guid.Parse("467ab87d-9494-4add-bf01-703540067b9e");
@@ -33,9 +35,7 @@ public class AuditLogger(
     private readonly IClock _clock = clock;
     private readonly IOutboxClient _outboxClient = outboxClient;
     private readonly IUnitOfWork _unitOfWork = unitOfWork;
-    private readonly IAuditUserContext _auditUserContext = auditUserContext
-                ?? throw new NullReferenceException("IAuditUserContext is required to use AuditLogger. " +
-                                                    "Did you register an implementation of it in the DI container?");
+    private readonly IUserContext<FrontendUser> _userContext = userContext;
 
     public async Task LogWithCommitAsync(
         AuditLogActivity activity,
@@ -44,7 +44,7 @@ public class AuditLogger(
         AuditLogEntityType? affectedEntityType,
         Guid? affectedEntityKey)
     {
-        var auditLogUser = _auditUserContext.CurrentUser;
+        var currentUser = _userContext.CurrentUser;
 
         var payloadAsJson = payload switch
         {
@@ -61,11 +61,11 @@ public class AuditLogger(
             payload: payloadAsJson,
             logId: Guid.NewGuid(),
             occuredOn: _clock.GetCurrentInstant(),
-            userId: auditLogUser.UserId,
-            actorId: auditLogUser.ActorId,
-            actorNumber: auditLogUser.ActorNumber,
-            marketRoles: auditLogUser.MarketRoles,
-            permissions: auditLogUser.Permissions,
+            userId: currentUser.UserId,
+            actorId: currentUser.Actor.ActorId,
+            actorNumber: currentUser.Actor.ActorNumber,
+            marketRoles: currentUser.Actor.MarketRole.ToString(),
+            permissions: string.Join(", ", currentUser.Actor.Permissions),
             affectedEntityType: affectedEntityType?.Identifier,
             affectedEntityKey: affectedEntityKey?.ToString());
