+++ source/dotnet/wholesale-api/Common.Interfaces/Security/FrontendActorMarketRole.cs
@@ -12,11 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Calculations.Interfaces.AuditLog;
+namespace Energinet.DataHub.Wholesale.Common.Interfaces.Security;
 
-public record AuditLogUser(
-    Guid UserId,
-    Guid ActorId,
-    string ActorNumber,
-    string MarketRoles,
-    string Permissions);
+public enum FrontendActorMarketRole
+{
+    /// <summary>
+    /// Other is used when a user's actor has a valid market role, but the role is currently irrelevant.
+    /// </summary>
+    Other,
+    GridAccessProvider,
+    EnergySupplier,
+    SystemOperator,
+    DataHubAdministrator,
+}
