+++ source/dotnet/wholesale-api/Common.Interfaces/Security/FrontendUser.cs
@@ -12,6 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Common.Infrastructure.Security;
+namespace Energinet.DataHub.Wholesale.Common.Interfaces.Security;
 
 public sealed record FrontendUser(Guid UserId, bool MultiTenancy, FrontendActor Actor);
