+++ source/dotnet/wholesale-api/Orchestrations/Extensions/DependencyInjection/OutboxProcessorExtensions.cs
@@ -1,0 +1,34 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using Energinet.DataHub.Core.Outbox.Abstractions;
+using Energinet.DataHub.Core.Outbox.Extensions.DependencyInjection;
+using Energinet.DataHub.Wholesale.Calculations.Infrastructure.AuditLog;
+using Energinet.DataHub.Wholesale.Calculations.Infrastructure.Persistence;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Energinet.DataHub.Wholesale.Orchestrations.Extensions.DependencyInjection;
+
+public static class OutboxProcessorExtensions
+{
+    public static IServiceCollection AddOutboxProcessing(this IServiceCollection services)
+    {
+        services.AddOutboxProcessor<DatabaseContext>();
+
+        // Add required outbox publishers
+        services.AddTransient<IOutboxPublisher, AuditLogOutboxMessageV1Publisher>();
+
+        return services;
+    }
+}
